Accesses path modifications
